2014-01-12 Roman - !!!!:SYS: Removed separated start scripts "openscada_demo", "openscada_AGLKS", "openscada_Boiler" and used instead         appropriated symbol links to "openscada_start".     Mostly build warnings resolved and some hiden bugs fixed.     FIX:Debug labels disable processing is fixed. - FIX:Archive.FSArch: Closing messages file on write error is fixed. - !:DB:DBF: Mostly code clean. - DAQ.{AMRDevs,BFN,Comedi}: Integer type expanded next to int64_t. - FIX:DAQ.SMH2Gi: AO forming and processing is fixed. - Special.FLibSYS: Object IO tested and adapted for main functions. - !:FIX:Special.SystemTests: Tests execution rules from config file fix.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2035 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/AMRDevs/da_Ergomera.cpp
@@ -115,9 +115,9 @@ void Ergomera::regVal( int reg )
 	workCnt.insert(workCnt.begin()+i_b,SDataRec(reg*2,2));
 }
 
-int Ergomera::getValR( int addr, ResString &err )
+int64_t Ergomera::getValR( int addr, ResString &err )
 {
-    int rez = EVAL_INT;
+    int64_t rez = EVAL_INT;
     vector< SDataRec >	&workCnt = acqBlks;
     for(unsigned i_b = 0; i_b < workCnt.size(); i_b++)
 	if((addr*2) >= workCnt[i_b].off && (addr*2+2) <= (workCnt[i_b].off+(int)workCnt[i_b].val.size()))
